+++ apps/main/public/design-system.html
@@ -737,7 +737,7 @@
               >variant="h4"</code
             >
             <div class="type-usage">
-              Section headers (ContactSection, LogoGrid, ScrollImageTextBlocks,
+              Section headers (CenteredTextSection, LogoGrid, ScrollImageTextBlocks,
               TwoColumnInterstitial)
             </div>
           </div>
